Reintroduced the original LRU policy for comparison, updating the file structure as necessary. Also added a new run script for convenience, and committed the trace I had been using for testing.
