+++ champsim_config.json
@@ -156,7 +156,7 @@
     "virtual_prefetch": false,
     "prefetch_activate": "LOAD,PREFETCH",
     "prefetcher": "no",
-    "replacement": "lru"
+    "replacement": "CARE"
   },
 
   "physical_memory": {
