+++ run.sh
@@ -1,0 +1,4 @@
+#bin/champsim --warmup-instructions 2000000 --simulation-instructions 5000000 649.fotonik3d_s-1B.champsimtrace.xz
+bin/champsim --warmup-instructions 200000 \
+--simulation-instructions 500000 \
+649.fotonik3d_s-1B.champsimtrace.xz
